adjusting setup.py and renaming folders with code
