+++ src/mcnntune/nnmodel.py
