+++ src/mcnntune/scripts/mcnntemplate.py
