+++ src/mcnntune/scripts/mcnntune.py
@@ -3,7 +3,7 @@
 Generates custom runcards for tunes variations
 """
 
-import mcnntunelib.app as vp
+import mcnntune.app as vp
 
 __author__ = "Stefano Carrazza & Simone Alioli"
 __version__= "1.0.0"
